Add automated test for the docker-build squash property

+++ tests/projects/docker-build/test.sh
@@ -11,16 +11,21 @@ testDockerBuild () {
   testName=docker-build
   testDir=$testsDir/docker-build
   printf "testing %s... " "$testName"
+  
+  ##########################################################
+  # Test1: Test docker-build by running the build pipeline #
+  ##########################################################
+
   # this test will create an image whose repository setting has the following suffix - should match the image-name setting in wercker.yml
-  repo=docker-build-image-name-1
+  imagename=docker-build-image-name-1
   # stop any existing container started by a previous run
   docker kill ${testName}-container > /dev/null 2>&1
   # delete any existing image built by a previous run
-  docker images | grep $repo | awk '{print $3}' | xargs -n1 docker rmi -f > /dev/null 2>&1
+  docker images | grep $imagename | awk '{print $3}' | xargs -n1 docker rmi -f > /dev/null 2>&1
   # check no existing image with the specified repository (column 1 is the repository)
-  docker images | awk '{print $1}' | grep -q "$repo"
+  docker images | awk '{print $1}' | grep -q "$imagename"
   if [ $? -eq 0 ]; then
-    echo "An image with repository $repo already exists"
+    echo "An image with repository $imagename already exists"
     return 1
   fi
   # now run the build pipeline - this creates an image with the specified repository setting
@@ -33,13 +38,20 @@ testDockerBuild () {
     return 1
   fi
   # verify that an image was created with the expected repository setting (column 1 is the repository)
-  docker images | awk '{print $1}' | grep -q "$repo"
+  docker images | awk '{print $1}' | grep -q "$imagename"
   if [ $? -ne 0 ]; then
-    echo "An image with repository $repo was not found"
+    echo "An image with repository $imagename was not found"
     return 1
   fi
   # find the imageid of this image (we don't know the complete repository name as it is prefixed with build-id)
-  imageid=`docker images | grep "$repo" | awk '{print $3}'`
+  imageid=`docker images | grep "$imagename" | awk '{print $3}'`
+  # verify that the image has NOT been squashed by doing "docker history" and checking that the most recent layer was NOT a merge
+  docker history $imageid  --format "{{.Comment}}" | head -1 | grep -q merge
+  if [ $? -eq 0 ]; then
+    echo "Image was unexpectedly merged"
+    return 1
+  fi
+
   # start the image using the docker CLI
   docker run --name ${testName}-container --rm -d -p 5000:5000 ${imageid} >> "${workingDir}/${testName}.log" 2>&1
   # test the image
@@ -49,7 +61,7 @@ testDockerBuild () {
   # stop the container
   docker kill ${testName}-container >> "${workingDir}/${testName}.log" 2>&1
   # delete the image we've just created
-  docker images | grep $repo | awk '{print $3}' | xargs -n1 docker rmi -f >> "${workingDir}/${testName}.log" 2>&1
+  docker images | grep $imagename | awk '{print $3}' | xargs -n1 docker rmi -f >> "${workingDir}/${testName}.log" 2>&1
   # now the container and image have been cleaned up, check whether the test worked
   if [ "$curlOutput1" != "Hello World!" ]; then
     cat "${workingDir}/${testName}.log"
@@ -66,6 +78,39 @@ testDockerBuild () {
     echo "Unexpected response from test container for localhost:5000/env/bar " $curlOutput3
     return 1
   fi    
+
+  ##############################################################################
+  # Test2: test the "squash" property by running the build-and-squash pipeline #
+  ##############################################################################
+
+  # this test will create an image whose repository setting has the following suffix - should match the image-name setting in wercker.yml
+  imagename2=docker-build-image-name-2
+  # delete any existing image built by a previous run
+  docker images | grep $imagename2 | awk '{print $3}' | xargs -n1 docker rmi -f > /dev/null 2>&1
+  # check no existing image with the specified repository (column 1 is the repository)
+  docker images | awk '{print $1}' | grep -q "$imagename2"
+  if [ $? -eq 0 ]; then
+    echo "An image with repository $imagename2 already exists"
+    return 1
+  fi
+  # now run the build-and-squash pipeline to build a squashed image
+  $wercker build "$testDir" --pipeline build-and-squash --docker-local --working-dir "$workingDir" &> "${workingDir}/${testName}.log"
+  if [ $? -ne 0 ]; then
+    printf "failed\n"
+    if [ "${workingDir}/${testName}.log" ]; then
+      cat "${workingDir}/${testName}.log"
+    fi
+    return 1
+  fi
+  # find the imageid of this image (we don't know the complete repository name as it is prefixed with build-id)
+  imageid2=`docker images | grep $imagename2 | awk '{print $3}'`
+  # verify that the image HAS been squashed by doing "docker history" and checking that the most recent layer WAS a merge
+  docker history $imageid2 --format "{{.Comment}}" | head -1 | grep -q merge
+  if [ $? -ne 0 ]; then
+    echo "Image was not merged"
+    return 1
+  fi
+
   # test passed
   #cat "${workingDir}/${testName}.log"
   printf "passed\n"
